FASTQ formatted files

FASTQ formatted files

FASTQ is a text-based file format for representing DNA sequences along with qualities for each base. A FASTQ file stores a list of sequence records in the following format:

@{name} {description}?
{sequence}
+
{qualities}

Here is an example of one record from a FASTQ file:

@FSRRS4401BE7HA
tcagTTAAGATGGGAT
+
###EEEEEEEEE##E#

Readers and Writers

The reader and writer for FASTQ formatted files, are found within the BioSequences.FASTQ module.

FASTX.FASTQ.Reader — Type.

FASTQ.Reader(input::IO; fill_ambiguous=nothing)

Create a data reader of the FASTQ file format.

Arguments

input: data source

fill_ambiguous=nothing: fill ambiguous symbols with the given symbol

source

FASTX.FASTQ.Writer — Type.

FASTQ.Writer(output::IO; quality_header=false)

Create a data writer of the FASTQ file format.

Arguments

output: data sink

quality_header=false: output the title line at the third line just after '+'

source

They can be created with IOStreams:

using FASTX

r = FASTQ.Reader(open("my-reads.fastq", "r"))
w = FASTQ.Writer(open("my-output.fastq", "w"))

Alternatively, Base.open is overloaded with a method for conveinience:

r = open(FASTQ.Reader, "/my-reads.fastq")
w = open(FASTQ.Writer, "my-out.fastq")

Note that FASTQ.Reader does not support line-wraps within sequence and quality. Usually sequence records will be read sequentially from a file by iteration.

reader = open(FASTQ.Reader, "my-reads.fastq")
for record in reader
    ## Do something
end
close(reader)

You can also overwrite records in a while loop to avoid excessive memory allocation.

reader = open(FASTQ.Reader, "my-reads.fastq")
record = FASTQ.Record()
while !eof(reader)
    read!(reader, record)
    ## Do something.
end

Reading in a record from a FASTQ formatted file will give you a variable of type FASTQ.Record.

FASTX.FASTQ.Record — Type.

FASTQ.Record()

Create an unfilled FASTQ record.

source

FASTQ.Record(data::Vector{UInt8})

Create a FASTQ record object from data.

This function verifies and indexes fields for accessors.

Warning

Note that the ownership of data is transferred to a new record object. Editing the input data will edit the record, and is not advised after construction of the record.

source

FASTQ.Record(str::AbstractString)

Create a FASTQ record object from str.

This function verifies and indexes fields for accessors.

source

FASTQ.Record(identifier, sequence, quality; offset=33)

Create a FASTQ record from identifier, sequence and quality.

source

FASTQ.Record(identifier, description, sequence, quality; offset=33)

Create a FASTQ record from identifier, description, sequence and quality.

source

Various getters and setters are available for FASTQ.Records:

FASTX.FASTQ.hasidentifier — Function.

hasidentifier(record::Record)

Checks whether or not the record has an identifier.

source

FASTX.FASTQ.identifier — Function.

identifier(record::Record)::Union{String,Nothing}

Get the sequence identifier of record.

Note

Returns nothing if the record has no identifier.

source

FASTX.FASTQ.hasdescription — Function.

hasdescription(record::Record)

Checks whether or not the record has a description.

source

FASTX.FASTQ.description — Function.

description(record::Record)::Union{String, Nothing}

Get the description of record.

Note

Returns nothing if record has no description.

source

FASTX.FASTQ.hassequence — Function.

hassequence(record::Record)

Checks whether or not a sequence record contains a sequence.

Note

Zero-length sequences are allowed in records.

source

FASTX.FASTQ.sequence — Function.

sequence(::Type{S}, record::Record, [part::UnitRange{Int}])

Get the sequence of record.

S can be either a subtype of BioSequences.BioSequence or String. If part argument is given, it returns the specified part of the sequence.

source

sequence(::Type{String}, record::Record, [part::UnitRange{Int}])::String

Get the sequence of record as a String. If part argument is given, it returns the specified part of the sequence.

source

sequence(record::Record, [part::UnitRange{Int}])::BioSequences.DNASequence

Get the sequence of record.

source

FASTX.FASTQ.hasquality — Function.

hasquality(record::Record)

Check whether the given FASTQ record has a quality string.

source

FASTX.FASTQ.quality — Function.

quality(record::Record, [offset::Integer=33, [part::UnitRange]])::Vector{UInt8}

Get the base quality of record.

source

quality(record::Record, encoding_name::Symbol, [part::UnitRange])::Vector{UInt8}

Get the base quality of record by decoding with encoding_name.

The encoding_name can be either :sanger, :solexa, :illumina13, :illumina15, or :illumina18.

Note

Returns nothing if the record has no quality string.

source

To write a BioSequence to FASTQ file, you first have to create a FASTQ.Record:

FASTX.FASTQ.Record — Method.

FASTQ.Record(identifier, description, sequence, quality; offset=33)

Create a FASTQ record from identifier, description, sequence and quality.

source

As always with julia IO types, remember to close your file readers and writer after you are finished.

Using open with a do-block can help ensure you close a stream after you are finished.

open(FASTQ.Reader, "my-reads.fastq") do reader
    for record in reader
        ## Do something
    end
end

Quality encodings

FASTQ records have a quality string which have platform dependent encodings. The FASTQ submodule has encoding and decoding support for the following quality encodings. These can be used with a FASTQ.quality method, to ensure the correct quality score values are extracted from your FASTQ quality strings.

FASTX.FASTQ.SANGER_QUAL_ENCODING — Constant.

Sanger (Phred+33) quality score encoding

source

FASTX.FASTQ.SOLEXA_QUAL_ENCODING — Constant.

Solexa (Solexa+64) quality score encoding

source

FASTX.FASTQ.ILLUMINA13_QUAL_ENCODING — Constant.

Illumina 1.3 (Phred+64) quality score encoding

source

FASTX.FASTQ.ILLUMINA15_QUAL_ENCODING — Constant.

Illumina 1.5 (Phred+64) quality score encoding

source

FASTX.FASTQ.ILLUMINA18_QUAL_ENCODING — Constant.

Illumina 1.8 (Phred+33) quality score encoding

source